[22.10.06][add] NCP rules about ACG, LB, NACL, Encrpytion
