+++ checkov/terraform/checks/resource/ncp/AccessControlGroupInboundRulePort22.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.AccessControlGroupInboundRule import 
 
 class AccessControlGroupGRuleInboundPort22(AccessControlGroupInboundRule):
     def __init__(self):
-        super().__init__(check_id="CUSTOM_NCP_ACG_003", port=22)
+        super().__init__(check_id="CUSTOM_NCP_ACG_004", port=22)
 
 
 check = AccessControlGroupGRuleInboundPort22()
