+++ checkov/terraform/checks/resource/ncp/AccessControlGroupInboundRulePort3389.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.AccessControlGroupInboundRule import 
 
 class AccessControlGroupGRuleInboundPort3389(AccessControlGroupInboundRule):
     def __init__(self):
-        super().__init__(check_id="CUSTOM_NCP_ACG_004", port=3389)
+        super().__init__(check_id="CUSTOM_NCP_ACG_005", port=3389)
 
 
 check = AccessControlGroupGRuleInboundPort3389()
