+++ checkov/terraform/checks/resource/ncp/LaunchConfigurationEncryptionVPC.py
@@ -7,7 +7,7 @@ class LaunchConfigurationEncryptionVPC(BaseResourceCheck):
     def __init__(self):
         name = "You can set whether to encrypt basic block storage if server image is RHV. Default false." \
                " Only support VPC environment."
-        id = "CKV_NCP_4"
+        id = "CKV_NCP_7"
         supported_resources = ['ncloud_launch_configuration']
         categories = [CheckCategories.ENCRYPTION]
         super().__init__(name=name, id=id, categories=categories, supported_resources=supported_resources)
