+++ checkov/terraform/checks/resource/ncp/NACLIngress20.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.NACLIngressCheck import NACLIngressCh
 
 class NACLIngress20(NACLIngressCheck):
     def __init__(self):
-        super().__init__(check_id="CKV_NCP_230", port=20)
+        super().__init__(check_id="CKV_NCP_10", port=20)
 
 
 check = NACLIngress20()
