+++ checkov/terraform/checks/resource/ncp/NACLIngress21.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.NACLIngressCheck import NACLIngressCh
 
 class NACLIngress21(NACLIngressCheck):
     def __init__(self):
-        super().__init__(check_id="CKV_NCP_229", port=21)
+        super().__init__(check_id="CKV_NCP_11", port=21)
 
 
 check = NACLIngress21()
