+++ checkov/terraform/checks/resource/ncp/NACLIngress22.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.NACLIngressCheck import NACLIngressCh
 
 class NACLIngress22(NACLIngressCheck):
     def __init__(self):
-        super().__init__(check_id="CKV_NCP_232", port=22)
+        super().__init__(check_id="CKV_NCP_12", port=22)
 
 
 check = NACLIngress22()
