+++ checkov/terraform/checks/resource/ncp/NACLIngress3389.py
@@ -3,7 +3,7 @@ from checkov.terraform.checks.resource.ncp.NACLIngressCheck import NACLIngressCh
 
 class NACLIngress3389(NACLIngressCheck):
     def __init__(self):
-        super().__init__(check_id="CKV_NCP_231", port=3389)
+        super().__init__(check_id="CKV_NCP_13", port=3389)
 
 
 check = NACLIngress3389()
