+++ tests/terraform/checks/resource/ncp/example_AccessControlGroupEgressCheck/main.tf
