+++ tests/terraform/checks/resource/ncp/example_AccessControlGroupIngressCheck/main.tf
